requirements.txt
absl-py==2.4.0
astunparse==1.6.3
beautifulsoup4==4.14.3
bidict==0.23.1
blinker==1.9.0
cachelib==0.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
coolname==2.2.0
deepface==0.0.99
dlib @ https://github.com/z-mahmud22/Dlib_Windows_Python3.x/raw/main/dlib-19.24.99-cp312-cp312-win_amd64.whl#sha256=20c62e606ca4c9961305f7be3d03990380d3e6c17f8d27798996e97a73271862
dnspython==2.8.0
eventlet==0.41.0
filelock==3.25.2
fire==0.7.1
Flask==3.0.0
flask-cors==6.0.2
Flask-Mail==0.9.1
Flask-MySQLdb==2.0.0
Flask-Session==0.5.0
Flask-SocketIO==5.3.6
Flask-WTF==1.2.1
flatbuffers==25.12.19
gast==0.7.0
gdown==5.2.1
google-pasta==0.2.0
greenlet==3.4.0
grpcio==1.80.0
gunicorn==25.3.0
h11==0.16.0
h5py==3.14.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
keras==3.14.0
libclang==18.1.1
lightdsa==0.0.3
lightecc==0.0.5
lightphe==0.0.22
lz4==4.4.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
mtcnn==1.0.0
mysqlclient==2.2.8
namex==0.1.0
numpy==1.26.0
opencv-python==4.11.0.86
opencv-python-headless==4.8.1.78
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==2.1.1
pillow==12.2.0
protobuf==7.34.1
Pygments==2.20.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-engineio==4.13.1
python-socketio==5.16.1
pytz==2026.1.post1
requests==2.33.1
retina-face==0.0.17
rich==14.3.3
setuptools==82.0.1
simple-websocket==1.1.0
six==1.17.0
soupsieve==2.8.3
stripe==7.1.0
sympy==1.14.0
tensorflow==2.21.0
termcolor==3.3.0
tf_keras==2.21.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
Werkzeug==3.0.0
wget==3.2
wheel==0.46.3
wrapt==2.1.2
wsproto==1.3.2
WTForms==3.2.1
